Disable element-wise sum offload in BYOC
This commit disables element-wise sum offloading in BYOC,
and a handful of other offloaded operations.
The reason they are disabled is that we use a new style of off-
loading to the accelerator (ethos-u style) which aggregates a
set of supported/required operations to byoc codegen.cc

This commit also explicitely disables element-wise sum.
This is because the EWS was refactored in the hardware library
and since those changes are not yet taken into account in
codegen.cc triggering a segmentation fault.