Install dory in Dockerfile
This commit installs dory in the /dory folder of the container
Currently it uses `-e` installation option for pip, due to an
upstream dory issue.

When this issue is solved, the normal `pip install .` should
be usable.

Another issue is that we currently don't check for a dory version.
We should have stable Dory versioning so that we can correctly
implement this.