Add wrapper for malloc and make malloc different for each target
The previous commit did not use the proper L2 malloc for tvm with
pulprt, problem with the pulprt is that it does not have a
free implementation that can just ingest a pointer.
In the pulprt free implementation the size has to be made explicit.

This commit adds the malloc_wrapper which calls pulp_rt's malloc
but in doing so allocates 4 bytes extra (at the beginning of the
allocated chunck) to store the size of the allocation for later
use in free_wrapper.
free_wrapper wraps pulp_rt's free but extracts the size info
from the header that was allocated with malloc_wrapper.
As such the need to explicitly include size info is gone.

This was necessary to be able to implement the TVM runtime,
The TVM runtime does not explicitely state how many bytes are
freed with TVMBackendFreeWorkspace.

All the changes are integrated in utils.py so that
the wrappers are generated when targetting pulp and the
regular mallocs are generated when targetting x86.