Refactor malloc and free wrapper
Malloc and free wrappers originally developed for pulp runtime are
now also included in x86 builds, since this is better for testing.

The wrappers are now always used, but the -DPULP compilation
preprocessor option now toggles the right "wrapped" version of
malloc and free in src/malloc_wrapper.c

This implementation also allows users to wrap their malloc
on x86 and pulp platforms with minimal effort.
Since the selection of x86/PULP is now done by the C preprocessor.
(It has made the implementation of utils.py also a lot simpler)