Add wrapper for benchmark and gdb anchor
gdb anchor: serves as a relocatable breakpoint in optimized gcc
code. Because of the compiler attributes the function shouldn't
be inlined, relocated or optimized out.
This allows for repeatable benchmark experiments
The gdb anchor is also written in an architecture-independent
manner and thus can also be used in X86 code.
For this reason, the target-dependency for demo code generation
was removed in utils.py.

benchmark wrapper. Wraps performance counter code in no-optimize,
no-inline blocks, In this way GCC -O3 optimized can still be
profiled reliably.
furthermore, it reduces the binary size overhead of inserting
benchmarking related functionality.