Change default C device to arm_cpu
This commit changes the wrapper code for TVMC to always selsect
the arm_cpu device.

By default TVM selects the "cpu" device, which actually performs
x86 optimized tiling and loop reordering.
In practice this means that TVM will replace NCHW conv2d operations
with NCHWc operations, which requires an extra layout
transformation.

This is useful in the case of automatic vectorization, but not in
our case, since we have to disable automatic vectorization to
work with the uTVM runtime anyway.