Add test script for ResNet
The test script has to be performed on X86, inside the TVM-fork
tutorial container, since it requires a working version of the
pulp sdk.

The script has shown a discrepancy in the result when TVM's fusion
pass is turned off on X86, since there is an error in passing on
the result in a 16 bit buffer.

the result after the first res layer is -2 for all elements of the
input tensor, however, for some reason this value is passed on
incorrectly to the next layer when compiling the full network.

Turning the fusion pass off still allows for verification testing
so no fix is presented in this commit.
Please just use -O0 and no fusion when testing for now.