Add float support in compare_outputs, gdb_demo.sh and add gdb for pulp
This commmit automatically make a gdb_demo script based on the
compiled network.
This is necessary if the network outputs something different from
A regular int8 as output value
It also casts all read in values from a string to a float.
This is useful if the integer or float values need to be compared.

Furthermore this commit adds an option to use a locally installed
riscv gdb version, which can be used with remote debugging over
tcp:3333