Support multiple inputs in create_demo_file
This commit adds support for multiple inputs to be created in
utils.py::create_demo_file, by setting the no_of_inputs argument.

Note that this function argument does not do any sanity check.
This means that you can possible generate C code for inputs that
do not necessarily exist.

This change also requires that all networks inputs are numbered
from now on.
When declaring a relay.Var you should set the name of the inputs to
"input_x" where x is a number starting from 0 and increasing in
value for more inputs