Change test parametrization of conv2d and dwconv2d tests
This commit introduces a function `get_test_params` that
replaces a set of pytest.mark.parametrize fixture markers

The reason for this change is that this in some cases generates
very long paths in combination with the tmp_path fixture
which end up in the debug symbol table of the final binary.
This in turn has resulted in some pulp riscv gdb crashes.

To avoid these crashes this new test parametrization method merges
all different parameters into one variable and assigns a single
(short) test id to each test. The variables are again unpacked
inside the test.
In this way the filenames stay short, (and cryptic :( )
but GDB does not seem to crash anymore :)