Fix missing DMA transfer in addition kernel
This commit fixes a missing DMA transfer in the addition kernel.
The Dory template has a field called branch_out which can be un-
toggled to not perform L1->L2 memory transfers between layers that
would fit L1 memory across kernel calls.
In our current flow there is no way to optimize for this across
kernel calls, so this can not be used yet.
Right now the DMA transfer is always performed, regardless if this
is necessary or not.