Small change in onnx transform
If the division factor of divide (from div -> floor -> max -> min
pattern) is 1.0, the onnx optimizer (called in quantlib during onnx export),
eliminates the division op (its configured to do basic optimizations).
This is not wanted but there is no way to control this specific
optimization. If optimizations are off, the graph contains too much
clutter.
To fix this, we modified the onnx transform in tvm to also consider
floor -> max -> min patterns without divide, as valid requantization patterns.

Tested on ResNet8