Small improvement layout transform
Now, we iterate on optimising the transforms until the optimization
saturates, which is much better than a fixed amount of iterations.
This reduces the amount of tfms for resnet20 with 1 more compared to the
previous code.
Also ensured that InferTypes is ran before pattern matching for
composite ops, since the pattern matcher requires checked types.